Move fluxtables to dygraph into a worker job
